Fix .stbad header: store tickSize as double (match 2.0.0 double-ticksize)
The 2.0.0 release changed Instrument.TickSize/PointValue from decimal to double.
The .stbad encoder now writes an 8-byte double for tickSize, but StbadHeader still
read a 16-byte decimal — corrupting the header on decode. Read it as a double to
match. Re-tested: 431/431 pass.

+++ src/SquidEyes.Pricing/Stbad/StbadHeader.cs
@@ -14,7 +14,7 @@ internal sealed class StbadHeader
     public required Source Source { get; init; }
     public required SessionKind Session { get; init; }
     public required bool HasOrderCount { get; init; }
-    public required decimal TickSize { get; init; }
+    public required double TickSize { get; init; }
     public required double PointValue { get; init; }
     public required DateTime SessionStart { get; init; }
     public required byte Codec { get; init; }
@@ -54,7 +54,7 @@ internal sealed class StbadHeader
 
         var fields = r.ReadByte();
         var hasCount = (fields & StbadFormat.FieldHasOrderCount) != 0;
-        var tickSize = r.ReadDecimal();
+        var tickSize = r.ReadDouble();
         var pointValue = r.ReadDouble();
 
         var tsResolution = r.ReadByte();
